prefer totalTermFrequency over docFreq in PhraseSuggester

+++ src/main/java/org/elasticsearch/search/suggest/phrase/LinearInterpoatingScorer.java
@@ -44,7 +44,7 @@ public final class LinearInterpoatingScorer extends WordScorer {
     @Override
     protected double scoreBigram(Candidate word, Candidate w_1) throws IOException {
         SuggestUtils.join(separator, spare, w_1.term, word.term);
-        final int count = frequency(spare);
+        final long count = frequency(spare);
         if (count < 1) {
             return unigramLambda * scoreUnigram(word);
         }
@@ -54,7 +54,7 @@ public final class LinearInterpoatingScorer extends WordScorer {
     @Override
     protected double scoreTrigram(Candidate w, Candidate w_1, Candidate w_2) throws IOException {
         SuggestUtils.join(separator, spare, w.term, w_1.term, w_2.term);
-        final int count = frequency(spare);
+        final long count = frequency(spare);
         if (count < 1) {
             return scoreBigram(w, w_1);
         }
